+++ ChangeLog.markdown
@@ -1,6 +1,14 @@
 # Yet Another DataTables Column Filter (Yadcf) ChangeLog
 
 
+## 0.3.3
+
+* Added two options:
+   - sort_as : Defines how the values in the filter will be sorted, alphabetically or numerically (default is alphabetically)
+   - sort_order : Defines the order in which the values in the filter will be sorted, ascending or descending  (default is ascending)
+
+
+
 ## 0.3.1
 
 * Better align of filter + close button (inside a wrapper div)